// web/src/openapi/config/iscsi.ts
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

/**
 * iSCSI config.
 */
export interface Config {
  /**
   * Initiator name.
   */
  initiator?: string;
  /**
   * List of targets.
   */
  targets?: Target[];
}
export interface Target {
  /**
   * IP address.
   */
  address: string;
  port: number;
  name: string;
  interface: string;
  startup?: "onboot" | "manual" | "automatic";
  authByTarget?: Authentication;
  authByInitiator?: Authentication;
}
export interface Authentication {
  username: string;
  password: string;
}